Added property aggregate
