+++ NextGen-BM-BE/NextGen-BM-BE-Domain/Entities/PropertyAggregate/PropertyExpense.cs
@@ -1,4 +1,4 @@
-namespace NextGen_BM_BE_Domain.Entities{
+namespace NextGen_BM_BE_Domain.Entities.PropertyAggregate{
 public class PropertyExpense{
     public int PropertyExpenseID { get; set; }
     public int PropertyExpenseTemplateID { get; set; }
@@ -6,6 +6,6 @@ public class PropertyExpense{
     public double Price { get; set; }
     public DateOnly StartDate { get; set; }
     public DateOnly EndDate { get; set; }
-    public string Description { get; set; }
+    public string? Description { get; set; }
 }
 };
