+++ NextGen-BM-BE/NextGen-BM-BE-Domain/Entities/PropertyAggregate/Specifications/PropertySpecifications.cs
@@ -1,0 +1,16 @@
+using Ardalis.Specification;
+
+namespace NextGen_BM_BE_Domain.Entities.PropertyAggregate.Specifications{
+
+    public class PropertySpecifications: Specification<Property>
+    {
+        public PropertySpecifications(int propertyID)
+        {
+            Query.Where(property => property.PropertyID == propertyID)
+            .Include(property => property.Users)
+            .Include(property => property.Expenses)
+            .Include(property => property.Payments)
+            .Include(property => property.ResidentsHistory);
+        }
+    }
+};
